+++ fabric/src/main/java/com/jahirtrap/walljump/network/PacketHandler.java
@@ -1,9 +1,7 @@
 package com.jahirtrap.walljump.network;
 
 import com.jahirtrap.walljump.network.message.MessageFallDistance;
-import com.jahirtrap.walljump.network.message.MessageServerConfig;
 import com.jahirtrap.walljump.network.message.MessageWallJump;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 
 public class PacketHandler {
@@ -12,7 +10,4 @@ public class PacketHandler {
         ServerPlayNetworking.registerGlobalReceiver(MessageWallJump.ID, MessageWallJump::handle);
     }
 
-    public static void initClient() {
-        ClientPlayNetworking.registerGlobalReceiver(MessageServerConfig.ID, MessageServerConfig::handle);
-    }
 }