+++ fabric/src/main/java/com/jahirtrap/walljump/network/message/MessageWallJump.java
@@ -1,6 +1,6 @@
 package com.jahirtrap.walljump.network.message;
 
-import com.jahirtrap.walljump.init.ServerConfig;
+import com.jahirtrap.walljump.init.ModConfig;
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
@@ -18,7 +18,7 @@ public class MessageWallJump {
         server.execute(() -> {
             if (didWallJump) {
                 player.resetFallDistance();
-                player.causeFoodExhaustion((float) ServerConfig.exhaustionWallJump);
+                player.causeFoodExhaustion((float) ModConfig.exhaustionWallJump);
             }
         });
     }