+++ CHANGELOG.md
@@ -1,1 +1,2 @@
+- Fixed server join failing with "registry data" errors when enchantments are disabled in the config ([Issue #37](https://github.com/jahirxtrap/walljump/issues/37))
 - Updated config lib